Production fixes

+++ flashcard-app/db/schema.rb
@@ -16,13 +16,13 @@ ActiveRecord::Schema[8.0].define(version: 2025_09_14_182651) do
     t.string "answer"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
-    t.integer "user_id", null: false
+    t.bigint "user_id", null: false
     t.integer "bucket"
     t.index ["user_id"], name: "index_flashcards_on_user_id"
   end
 
   create_table "sessions", force: :cascade do |t|
-    t.integer "user_id", null: false
+    t.bigint "user_id", null: false
     t.string "ip_address"
     t.string "user_agent"
     t.datetime "created_at", null: false
